Split the implementation of RemoteImageBufferProxy and RemoteImageBuffer to source and header files https://bugs.webkit.org/show_bug.cgi?id=242835
Reviewed by Simon Fraser.

After 252266@main, it is now possible to split the implementation of RemoteImageBuffer
and RemoteImageBufferProxy to source and header files since they are not templates
anymore.

* Source/WebCore/page/Chrome.cpp:
(WebCore::Chrome::createImageBuffer const):
* Source/WebCore/page/Chrome.h:
* Source/WebCore/page/ChromeClient.h:
(WebCore::ChromeClient::createImageBuffer const):
* Source/WebCore/platform/HostWindow.h:
* Source/WebKit/GPUProcess/graphics/RemoteImageBuffer.cpp: Added.
(WebKit::RemoteImageBuffer::RemoteImageBuffer):
(WebKit::m_renderingResourcesRequest):
(WebKit::RemoteImageBuffer::~RemoteImageBuffer):
(WebKit::RemoteImageBuffer::setOwnershipIdentity):
* Source/WebKit/GPUProcess/graphics/RemoteImageBuffer.h:
(WebKit::RemoteImageBuffer::RemoteImageBuffer): Deleted.
(WebKit::RemoteImageBuffer::m_renderingResourcesRequest): Deleted.
(WebKit::RemoteImageBuffer::~RemoteImageBuffer): Deleted.
(WebKit::RemoteImageBuffer::setOwnershipIdentity): Deleted.
* Source/WebKit/GPUProcess/graphics/RemoteRenderingBackend.cpp:
* Source/WebKit/GPUProcess/media/RemoteLegacyCDMSessionProxy.cpp:
* Source/WebKit/Sources.txt:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.cpp: Added.
(WebKit::RemoteImageBufferProxy::RemoteImageBufferProxy):
(WebKit::RemoteImageBufferProxy::~RemoteImageBufferProxy):
(WebKit::RemoteImageBufferProxy::waitForDidFlushOnSecondaryThread):
(WebKit::RemoteImageBufferProxy::didFlush):
(WebKit::RemoteImageBufferProxy::backingStoreWillChange):
(WebKit::RemoteImageBufferProxy::waitForDidFlushWithTimeout):
(WebKit::RemoteImageBufferProxy::ensureBackendCreated const):
(WebKit::RemoteImageBufferProxy::copyNativeImage const):
(WebKit::RemoteImageBufferProxy::drawConsuming):
(WebKit::RemoteImageBufferProxy::sinkIntoNativeImage):
(WebKit::RemoteImageBufferProxy::filteredImage):
(WebKit::RemoteImageBufferProxy::getPixelBuffer const):
(WebKit::RemoteImageBufferProxy::clearBackend):
(WebKit::RemoteImageBufferProxy::context const):
(WebKit::RemoteImageBufferProxy::drawingContext):
(WebKit::RemoteImageBufferProxy::putPixelBuffer):
(WebKit::RemoteImageBufferProxy::convertToLuminanceMask):
(WebKit::RemoteImageBufferProxy::transformToColorSpace):
(WebKit::RemoteImageBufferProxy::flushContext):
(WebKit::RemoteImageBufferProxy::flushDrawingContext):
(WebKit::RemoteImageBufferProxy::flushDrawingContextAsync):
(WebKit::RemoteImageBufferProxy::createFlusher):
(WebKit::RemoteImageBufferProxy::prepareForBackingStoreChange):
* Source/WebKit/WebProcess/GPU/graphics/RemoteImageBufferProxy.h:
(WebKit::RemoteImageBufferProxy::create):
(WebKit::RemoteImageBufferProxy::putPixelBuffer):
(WebKit::RemoteImageBufferProxy::~RemoteImageBufferProxy): Deleted.
(WebKit::RemoteImageBufferProxy::waitForDidFlushOnSecondaryThread): Deleted.
(WebKit::RemoteImageBufferProxy::RemoteImageBufferProxy): Deleted.
(WebKit::RemoteImageBufferProxy::waitForDidFlushWithTimeout): Deleted.
(WebKit::RemoteImageBufferProxy::recordNativeImageUse): Deleted.
(WebKit::RemoteImageBufferProxy::recordFontUse): Deleted.
(WebKit::RemoteImageBufferProxy::recordImageBufferUse): Deleted.
(WebKit::RemoteImageBufferProxy::prepareForBackingStoreChange): Deleted.
* Source/WebKit/WebProcess/GPU/graphics/RemoteRenderingBackendProxy.cpp:
(WebKit::RemoteRenderingBackendProxy::createImageBuffer):
* Source/WebKit/WebProcess/GPU/graphics/RemoteRenderingBackendProxy.h:
* Source/WebKit/WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::createImageBuffer const):
* Source/WebKit/WebProcess/WebCoreSupport/WebChromeClient.h:

Canonical link: https://commits.webkit.org/252577@main
