+++ Source/WebKit/GPUProcess/media/RemoteLegacyCDMSessionProxy.cpp
@@ -32,6 +32,7 @@
 #include "RemoteLegacyCDMFactoryProxy.h"
 #include "RemoteLegacyCDMSessionMessages.h"
 #include <JavaScriptCore/GenericTypedArrayViewInlines.h>
+#include <JavaScriptCore/TypedArrayAdaptors.h>
 #include <WebCore/LegacyCDM.h>
 #include <WebCore/SharedBuffer.h>
 #include <wtf/LoggerHelper.h>
